services/ingestion/requirements.txt
# services/ingestion/requirements.txt
qdrant-client==1.7.3
python-dotenv==1.0.0
httpx==0.25.2
numpy==1.26.0
loguru==0.7.2
pydantic==2.5.2
sentence-transformers==2.2.2
apscheduler==3.10.4
tenacity==8.2.3